operator/requirements.txt
flask>=3.0.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
werkzeug>=3.0.0
apscheduler>=3.10.4
twilio>=9.0.0
requests>=2.31.0
